+++ Assets/OkapiKit/Scripts/Systems/Path.cs
@@ -10,10 +10,10 @@ namespace OkapiKit
     [AddComponentMenu("Okapi/Other/Path Object")]
     public class Path : OkapiElement
     {
-        [SerializeField] public enum PathType { Linear = 0, Smooth = 1, Circle = 2, Arc = 3, Polygon = 4};
+        [SerializeField] public enum Type { Linear = 0, Smooth = 1, Circle = 2, Arc = 3, Polygon = 4};
 
         [SerializeField]
-        private PathType    type = PathType.Linear;
+        private Type    type = Type.Linear;
         [SerializeField]
         private bool        closed = false;
         [SerializeField, Range(3, 30)]
@@ -31,7 +31,7 @@ namespace OkapiKit
         [SerializeField]
         private Color       displayColor = Color.white;
 
-        private bool isSmooth => type == PathType.Smooth;
+        private bool isSmooth => type == Type.Smooth;
 
         private float           primaryRadius;
         private Vector2         primaryDir;
@@ -54,8 +54,8 @@ namespace OkapiKit
         }
         public int GetEditPointsCount() => points.Count;
 
-        public PathType GetPathType() => type;
-        public void SetPathType(PathType t) => type = t;
+        public Type GetPathType() => type;
+        public void SetPathType(Type t) => type = t;
         public void SetWorldSpace(bool ws) => worldSpace = ws;
 
         public bool isEditMode => editMode;
@@ -111,6 +111,20 @@ namespace OkapiKit
             return a * t3 + b * t2 + c * t + d;
         }
 
+        private static Vector3 EvaluateCatmullRomDerivative(Vector3 p0, Vector3 p1, Vector3 p2, Vector3 p3, float t)
+        {
+            float t2 = t * t;
+
+            Vector3 a = -0.5f * p0 + 1.5f * p1 - 1.5f * p2 + 0.5f * p3;
+            Vector3 b = p0 - 2.5f * p1 + 2f * p2 - 0.5f * p3;
+            Vector3 c = -0.5f * p0 + 0.5f * p2;
+
+            // Derivative calculation
+            Vector3 derivative = 3 * a * t2 + 2 * b * t + c;
+
+            return derivative;
+        }
+
         public Vector2 EvaluateWorld(float t)
         {
             return transform.TransformPoint(EvaluateLocal(t));
@@ -120,7 +134,7 @@ namespace OkapiKit
         {
             switch (type)
             {
-                case PathType.Linear:
+                case Type.Linear:
                     {
                         float d = t * linearLength;
                         
@@ -137,7 +151,7 @@ namespace OkapiKit
 
                         return points[0];
                     }
-                case PathType.Smooth:
+                case Type.Smooth:
                     {
                         int numPoints = points.Count;
                         int lastIndex = closed ? numPoints : numPoints - 2;
@@ -157,14 +171,14 @@ namespace OkapiKit
 
                         return EvaluateCatmullRom(p0, p1, p2, p3, t);
                     }
-                case PathType.Circle:
+                case Type.Circle:
                     {
                         float   angle = t * Mathf.PI * 2.0f;
                         Vector2 center = points[0];
 
                         return center + primaryRadius * primaryDir * Mathf.Cos(angle) + perpRadius * perpDir * Mathf.Sin(angle);
                     }
-                case PathType.Arc:
+                case Type.Arc:
                     {
                         float angle = Mathf.Lerp(startAngle, endAngle, t);
                         float radius = Mathf.Lerp(primaryRadius, secondaryRadius, t);
@@ -172,7 +186,7 @@ namespace OkapiKit
 
                         return center + Vector2.right * radius * Mathf.Cos(angle) + Vector2.up * radius * Mathf.Sin(angle);
                     }
-                case PathType.Polygon:
+                case Type.Polygon:
                     { 
                         float   totalSides = (nSides - ((closed) ? (0) : (1)));
                         int     side = Mathf.FloorToInt(totalSides * t);
@@ -195,6 +209,158 @@ namespace OkapiKit
             return Vector2.zero;
         }
 
+        public Vector2 EvaluateWorldDir(float t)
+        {
+            return transform.TransformVector(EvaluateLocalDir(t));
+        }
+
+        public Vector2 EvaluateLocalDir(float t)
+        {
+            switch (type)
+            {
+                case Type.Linear:
+                    {
+                        float d = t * linearLength;
+
+                        for (int i = 0; i < points.Count - ((closed) ? (0) : (1)); i++)
+                        {
+                            var delta = points[(i + 1) % points.Count] - points[i];
+                            float thisDistance = delta.magnitude;
+                            if (thisDistance >= d)
+                            {
+                                return delta / thisDistance;
+                            }
+                            else d -= thisDistance; 
+                        }
+
+                        if (points.Count > 2)
+                        {
+                            if (closed)
+                            {
+                                var delta = points[0] - points[points.Count - 1];
+                                return delta.normalized;
+                            }
+                            else
+                            {
+                                var delta = points[points.Count - 1] - points[points.Count - 2];
+                                return delta.normalized;
+                            }
+                        }
+
+                        return points[0];
+                    }
+                case Type.Smooth:
+                    {
+                        int numPoints = points.Count;
+                        int lastIndex = closed ? numPoints : numPoints - 2;
+                        float scaledT = t * lastIndex; // Scale t to the number of segments
+                        int i = Mathf.FloorToInt(scaledT) % numPoints; // Segment index, wrap around for closed curve
+
+                        // Calculate t for the specific segment
+                        t = scaledT - Mathf.FloorToInt(scaledT);
+
+                        // The three points for the segment
+                        Vector3 p0 = points[(i - 1 + numPoints) % numPoints];
+                        Vector3 p1 = points[i];
+                        Vector3 p2 = points[(i + 1) % numPoints];
+
+                        // Generate a fake fourth point for Catmull-Rom equation
+                        Vector3 p3 = points[(i + 2) % numPoints];
+
+                        return EvaluateCatmullRomDerivative(p0, p1, p2, p3, t).normalized;
+                    }
+                case Type.Circle:
+                    {
+                        float angle = t * Mathf.PI * 2.0f;
+
+                        return (-primaryDir * Mathf.Sin(angle) + perpDir * Mathf.Cos(angle)).normalized;
+                    }
+                case Type.Arc:
+                    {
+                        float angle = Mathf.Lerp(startAngle, endAngle, t);
+                        float radius = Mathf.Lerp(primaryRadius, secondaryRadius, t);
+
+                        return (Vector2.up * radius * Mathf.Cos(angle) - Vector2.right * radius * Mathf.Sin(angle)).normalized;
+                    }
+                case Type.Polygon:
+                    {
+                        float totalSides = (nSides - ((closed) ? (0) : (1)));
+                        int side = Mathf.FloorToInt(totalSides * t);
+                        float tInc = 1.0f / totalSides;
+                        float angleInc = Mathf.PI * 2.0f / (totalSides + ((closed) ? (0) : (1)));
+                        float angle = side * angleInc;
+                        Vector2 center = points[0];
+
+                        Vector3 p1 = center + primaryRadius * primaryDir * Mathf.Cos(angle) + perpRadius * perpDir * Mathf.Sin(angle);
+                        Vector3 p2 = center + primaryRadius * primaryDir * Mathf.Cos(angle + angleInc) + perpRadius * perpDir * Mathf.Sin(angle + angleInc);
+
+                        float remainingT = (t - side * tInc) / tInc;
+
+                        return (p2 - p1).normalized;
+                    }
+                default:
+                    break;
+            }
+
+            return Vector2.zero;
+        }
+
+        public void InvertPath()
+        {
+            if ((type == Path.Type.Linear) || (type == Path.Type.Smooth))
+            {
+                var newPoints = new List<Vector3>();
+                for (int i = points.Count - 1; i >= 0; i--)
+                {
+                    newPoints.Add(points[i]);
+                }
+
+                points = newPoints;
+            }
+            else if (type == Path.Type.Arc)
+            {
+                if (points.Count > 1)
+                {
+                    (points[1], points[2]) = (points[2], points[1]);
+                }
+            }
+        }
+
+        public void CenterPath()
+        {
+            if ((points == null) || (points.Count == 0)) return;
+
+            Vector3 delta = Vector3.zero;
+
+            if ((type == Path.Type.Linear) || (type == Path.Type.Smooth))
+            {
+                // Get extents of object
+                Vector3 min = points[0];
+                Vector3 max = points[0];
+
+                foreach (var pt in points)
+                {
+                    min.x = Mathf.Min(min.x, pt.x);
+                    min.y = Mathf.Min(min.y, pt.y);
+                    min.z = Mathf.Min(min.z, pt.z);
+                    max.x = Mathf.Max(max.x, pt.x);
+                    max.y = Mathf.Max(max.y, pt.y);
+                    max.z = Mathf.Max(max.z, pt.z);
+                }
+
+                delta = new Vector3(-(max.x + min.x) * 0.5f, -(max.y + min.y) * 0.5f, -(max.z + min.z) * 0.5f);
+            }
+            else if ((type == Path.Type.Arc) || (type == Path.Type.Circle) || (type == Path.Type.Polygon))
+            {
+                delta = -points[0];
+            }
+
+            for (int i = 0; i < points.Count; i++)
+            {
+                points[i] = points[i] + delta;
+            }
+        }
+
         public List<Vector3> GetPoints()
         {
             if ((!dirty) && (fullPoints == null)) return fullPoints;
@@ -207,12 +373,12 @@ namespace OkapiKit
             int steps = 20;
             switch (type)
             {
-                case PathType.Linear:
+                case Type.Linear:
                     steps = 0;
                     fullPoints = new List<Vector3>(points);
                     if ((closed) && (points.Count > 1)) fullPoints.Add(points[0]);
                     break;
-                case PathType.Smooth:
+                case Type.Smooth:
                     if (points.Count < 3)
                     {
                         steps = 0;
@@ -223,10 +389,10 @@ namespace OkapiKit
                         steps = points.Count * 20;
                     }
                     break;
-                case PathType.Circle:
-                case PathType.Arc:
+                case Type.Circle:
+                case Type.Arc:
                     {
-                        if (points.Count < ((type == PathType.Circle) ? (2) : (3)))
+                        if (points.Count < ((type == Type.Circle) ? (2) : (3)))
                         {
                             steps = 0;
                             fullPoints = new List<Vector3>(points);
@@ -237,7 +403,7 @@ namespace OkapiKit
                         }
                     }
                     break;
-                case PathType.Polygon:
+                case Type.Polygon:
                     {
                         if (points.Count < 2)
                         {
@@ -312,15 +478,15 @@ namespace OkapiKit
 
             if ((points == null) || (points.Count == 0))
             {
-                if (type == PathType.Circle)
+                if (type == Type.Circle)
                 {
                     _logs.Add(new LogEntry(LogEntry.Type.Error, "No points defined: for a circle we need at least 2 points!", "For a circle we need at least 2 points (center and a radius)"));
                 }
-                else if (type == PathType.Arc)
+                else if (type == Type.Arc)
                 {
                     _logs.Add(new LogEntry(LogEntry.Type.Error, "No points defined: for an arc we need at least 3 points!", "For an arc we need at least 3 points (center and two positions for the radius and directions of the arc)"));
                 }
-                else if (type == PathType.Polygon)
+                else if (type == Type.Polygon)
                 {
                     _logs.Add(new LogEntry(LogEntry.Type.Error, "No points defined: for a polygon we need at least 2 points!", "For a polygon we need at least 2 points (center and a radius)"));
                 }
@@ -331,15 +497,15 @@ namespace OkapiKit
             }
             else
             {
-                if ((type == PathType.Circle) && (points.Count < 2))
+                if ((type == Type.Circle) && (points.Count < 2))
                     {
                     _logs.Add(new LogEntry(LogEntry.Type.Error, "Not enough points defined: for a circle we need at least 2 points!", "For a circle we need at least 2 points (center and a radius)"));
                 }
-                else if ((type == PathType.Arc) && (points.Count < 3))
+                else if ((type == Type.Arc) && (points.Count < 3))
                 {
                     _logs.Add(new LogEntry(LogEntry.Type.Error, "Not enough points defined: for an arc we need at least 3 points!", "For an arc we need at least 3 points (center and two positions for the radius and directions of the arc)"));
                 }
-                else if ((type == PathType.Polygon) && (points.Count < 2))
+                else if ((type == Type.Polygon) && (points.Count < 2))
                 {
                     _logs.Add(new LogEntry(LogEntry.Type.Error, "Not enough points defined: for a polygon we need at least 2 points!", "For a polygon we need at least 2 points (center and a radius)"));
                 }
@@ -408,6 +574,12 @@ namespace OkapiKit
                 }
             }
         }
+
+        public Vector3 upAxis => primaryDir;
+        public float   upExtent => primaryRadius;
+        public Vector3 rightAxis => perpDir;
+        public float rightExtent => perpRadius;
+
         public void SetDirty()
         {
             dirty = true;